Create Factories for all models
This huge rewrite simplified how models are loaded and has moved most of
the logic into one place, by creating layers of indirection that are the
same for all cases: compilers, benchmarks, machines.

The current format is to have Factories for all models which will handle
paths and checks and leave to the ModelLoader to do the actual loading.
This removes any model logic from the controller and brings a unified
syntax for all models.

It also renames the files to the main class names, adds a lot of TODO
lines, type&value checks and better exception control when loading a
module, a few of the TODOs that we had before.