Adding Manifest file
The manifest file dumps all fields on the three models: benchmark,
compiler and machine, as well as the user options provided by argparse
and the environmt variables available to the program.

There was also some clean up of variables, to make sure we dump the same
names for the foreseeable future.