Adding cpu/mem parameters, name/affinity detection
Assuming Linux and using lscpu, MachineFactory now can autodetect
everything it needs to in order to guess architecture, cache hierarchy,
cpu affinity list, etc.

If the machine name is not passed, or the name is actually an
architecture, ie a generic model, the detection will populate the
model's parameters on loading.