Adding taskset support
This is a simplified implementation that will iterate through the
affinity list to taskset to specific cores that the machine set as
preferred.

This will work for single threaded programs (picking the most stable
core in the machine) as well as multiple process.

It will NOT work for threads, which needs pthread affinity support:
 - http://man7.org/linux/man-pages/man3/pthread_setaffinity_np.3.html

Nor OpenMP, which needs omp affinity support:
 - http://pages.tacc.utexas.edu/~eijkhout/pcse/html/omp-affinity.html