Forcing taskset even when no affinity is passed
This makes sure, at the minimum, the benchmark runs on a core that is
not the first one. Also only picking up-to 'threads'
parallel processes, and not running the same benchmark on sequentially
different process unless it's multiple binaries at the same time.

All those changes are supposed to increase stability.