Adding support for multiple urls/executables
Some benchmarks (like SPEC, EEMBC, BLAS-Tester) have multiple binaries,
others (like BLAS-Tester) needs dependencies (OpenBLAS), so adding
support for multiple repositories, multiple builds, and execution of
multiple benchmarks is necessary.

This will tie up in the future with the multi-threaded execution that we
want to get for SPEC/EEMBC/CoreMark, but it's not there yet. However,
having a list of executables will make the change in the Executable much
easier.