Better error message for Perf failures
Being more informative on perf errors, both when it's not installed and
when CAP_SYS_ADMIN is higher than 2, also mentioning which file to
change to fix that.

Next step is to make that into a warning and run without perf, but
benchmark results expect a dictionary and normal execution only provides
plain text (no parsers). Need to think a bit more about that.